+++ app/src/main/java/io/keepcoding/discourse_android/Data/Client/Http/DiscourseApi.kt
@@ -1,6 +1,6 @@
 package io.keepcoding.discourse_android.Data.Client.Http
 
-import io.keepcoding.discourse_android.Data.Models.LatestTopicResponse
+import io.keepcoding.discourse_android.Data.Models.ResponseModels.LatestTopicResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Headers
