+++ app/src/main/java/io/keepcoding/discourse_android/Data/Models/AppModels/Poster.kt
@@ -1,0 +1,6 @@
+package io.keepcoding.discourse_android.Data.Models.AppModels
+
+data class Poster (
+        val username: String = "",
+        val URL: String = ""
+)
