+++ app/src/main/java/io/keepcoding/discourse_android/Data/Models/AppModels/TimeOffset.kt
@@ -1,0 +1,3 @@
+package io.keepcoding.discourse_android.Data.Models.AppModels
+
+data class TimeOffset(val amount: Int, val unit: Int)
