+++ app/src/main/java/io/keepcoding/discourse_android/UI/topics/CallBackTopicClick.kt
@@ -1,6 +1,7 @@
 package io.keepcoding.discourse_android.UI.topics
 
-import io.keepcoding.discourse_android.Data.Models.TopicItem
+import io.keepcoding.discourse_android.Data.Models.AppModels.TopicItem
+
 
 interface CallbackTopicClick {
     fun onItemClick(topic: TopicItem)
